Add files via upload
The pumpkin seeds dataset support vector machine model is a machine learning model that was created to classify pumpkin seeds. The model is based on support vector machines, which are a type of supervised learning algorithm. The model was trained on a dataset of pumpkin seeds, and it is able to accurately classify pumpkin seeds.